+++ packages/common-web/src/async.ts
@@ -152,24 +152,75 @@ export class AsyncBufferFullError extends Error {
   }
 }
 
-export const handleAllSettledErrors = (
+/**
+ * Utility function that behaves like {@link Promise.allSettled} but returns the
+ * same result as {@link Promise.all} in case every promise is fulfilled, and
+ * throws an {@link AggregateError} if there are more than one errors.
+ */
+export function allFulfilled<T extends readonly unknown[] | []>(
+  promises: T,
+): Promise<{ -readonly [P in keyof T]: Awaited<T[P]> }>
+export function allFulfilled<T>(
+  promises: Iterable<T | PromiseLike<T>>,
+): Promise<Awaited<T>[]>
+export function allFulfilled(
+  promises: Iterable<Promise<unknown>>,
+): Promise<unknown[]> {
+  return Promise.allSettled(promises).then(handleAllSettledErrors)
+}
+
+export function handleAllSettledErrors<
+  T extends readonly PromiseSettledResult<unknown>[] | [],
+>(
+  results: T,
+): {
+  -readonly [P in keyof T]: T[P] extends PromiseSettledResult<infer U>
+    ? U
+    : never
+}
+export function handleAllSettledErrors<T>(
+  results: PromiseSettledResult<T>[],
+): T[]
+export function handleAllSettledErrors(
   results: PromiseSettledResult<unknown>[],
-) => {
-  const errors = results.filter(isRejected).map((res) => res.reason)
+): unknown[] {
+  const errors = results.filter(isRejectedResult).map(extractReason)
   if (errors.length === 0) {
-    return
+    // No need to filter here, it is safe to assume that all promises are fulfilled
+    return (results as PromiseFulfilledResult<unknown>[]).map(extractValue)
   }
   if (errors.length === 1) {
     throw errors[0]
   }
   throw new AggregateError(
     errors,
-    'Multiple errors: ' + errors.map((err) => err?.message).join('\n'),
+    `Multiple errors: ${errors.map(stringifyReason).join('\n')}`,
   )
 }
 
-const isRejected = (
+export function isRejectedResult(
   result: PromiseSettledResult<unknown>,
-): result is PromiseRejectedResult => {
+): result is PromiseRejectedResult {
   return result.status === 'rejected'
+}
+
+function extractReason(result: PromiseRejectedResult): unknown {
+  return result.reason
+}
+
+export function isFulfilledResult<T>(
+  result: PromiseSettledResult<T>,
+): result is PromiseFulfilledResult<T> {
+  return result.status === 'fulfilled'
+}
+
+function extractValue<T>(result: PromiseFulfilledResult<T>): T {
+  return result.value
+}
+
+function stringifyReason(reason: unknown): string {
+  if (reason instanceof Error) {
+    return reason.message
+  }
+  return String(reason)
 }