+++ packages/lexicon/src/validators/formats.ts
@@ -47,6 +47,7 @@ export function atUri(path: string, value: string): ValidationResult {
       error: new ValidationError(`${path} must be a valid at-uri`),
     }
   }
+
   return { success: true, value }
 }
 
@@ -59,6 +60,7 @@ export function did(path: string, value: string): ValidationResult {
       error: new ValidationError(`${path} must be a valid did`),
     }
   }
+
   return { success: true, value }
 }
 
@@ -71,21 +73,24 @@ export function handle(path: string, value: string): ValidationResult {
       error: new ValidationError(`${path} must be a valid handle`),
     }
   }
+
   return { success: true, value }
 }
 
 export function atIdentifier(path: string, value: string): ValidationResult {
-  const isDid = did(path, value)
-  if (!isDid.success) {
-    const isHandle = handle(path, value)
-    if (!isHandle.success) {
-      return {
-        success: false,
-        error: new ValidationError(`${path} must be a valid did or a handle`),
-      }
-    }
+  // We can discriminate based on the "did:" prefix
+  if (value.startsWith('did:')) {
+    const didResult = did(path, value)
+    if (didResult.success) return didResult
+  } else {
+    const handleResult = handle(path, value)
+    if (handleResult.success) return handleResult
   }
-  return { success: true, value }
+
+  return {
+    success: false,
+    error: new ValidationError(`${path} must be a valid did or a handle`),
+  }
 }
 
 export function nsid(path: string, value: string): ValidationResult {