+++ packages/ozone/src/api/moderation/queryStatuses.ts
@@ -18,6 +18,11 @@ export default function (server: Server, ctx: AppContext) {
         reportedBefore,
         ignoreSubjects,
         lastReviewedBy,
+        hostingDeletedBefore,
+        hostingDeletedAfter,
+        hostingUpdatedBefore,
+        hostingUpdatedAfter,
+        hostingStatuses,
         sortDirection = 'desc',
         sortField = 'lastReportedAt',
         includeMuted = false,
@@ -28,6 +33,9 @@ export default function (server: Server, ctx: AppContext) {
         excludeTags = [],
         collections = [],
         subjectType,
+        queueCount,
+        queueIndex,
+        queueSeed,
       } = params
       const db = ctx.db
       const modService = ctx.modService(db)
@@ -42,6 +50,11 @@ export default function (server: Server, ctx: AppContext) {
         reportedAfter,
         reportedBefore,
         includeMuted,
+        hostingDeletedBefore,
+        hostingDeletedAfter,
+        hostingUpdatedBefore,
+        hostingUpdatedAfter,
+        hostingStatuses,
         onlyMuted,
         ignoreSubjects,
         sortDirection,
@@ -53,6 +66,9 @@ export default function (server: Server, ctx: AppContext) {
         excludeTags,
         collections,
         subjectType,
+        queueCount,
+        queueIndex,
+        queueSeed,
       })
       const subjectStatuses = results.statuses.map((status) =>
         modService.views.formatSubjectStatus(status),