+++ packages/ozone/src/lexicon/types/tools/ozone/moderation/queryStatuses.ts
@@ -10,6 +10,12 @@ import { HandlerAuth, HandlerPipeThrough } from '@atproto/xrpc-server'
 import * as ToolsOzoneModerationDefs from './defs'
 
 export interface QueryParams {
+  /** Number of queues being used by moderators. Subjects will be split among all queues. */
+  queueCount?: number
+  /** Index of the queue to fetch subjects from. Works only when queueCount value is specified. */
+  queueIndex?: number
+  /** A seeder to shuffle/balance the queue items. */
+  queueSeed?: string
   /** All subjects, or subjects from given 'collections' param, belonging to the account specified in the 'subject' param will be returned. */
   includeAllUserRecords?: boolean
   /** The subject to get the status for. */
@@ -22,6 +28,16 @@ export interface QueryParams {
   reportedBefore?: string
   /** Search subjects reviewed after a given timestamp */
   reviewedAfter?: string
+  /** Search subjects where the associated record/account was deleted after a given timestamp */
+  hostingDeletedAfter?: string
+  /** Search subjects where the associated record/account was deleted before a given timestamp */
+  hostingDeletedBefore?: string
+  /** Search subjects where the associated record/account was updated after a given timestamp */
+  hostingUpdatedAfter?: string
+  /** Search subjects where the associated record/account was updated before a given timestamp */
+  hostingUpdatedBefore?: string
+  /** Search subjects by the status of the associated record/account */
+  hostingStatuses?: string[]
   /** Search subjects reviewed before a given timestamp */
   reviewedBefore?: string
   /** By default, we don't include muted subjects in the results. Set this to true to include them. */