+++ packages/ozone/tests/server.test.ts
@@ -1,8 +1,7 @@
-import { AddressInfo } from 'net'
+import { TestNetwork, TestOzone } from '@atproto/dev-env'
 import express from 'express'
-import axios, { AxiosError } from 'axios'
-import { TestOzone, TestNetwork } from '@atproto/dev-env'
 import { handler as errorHandler } from '../src/error'
+import { startServer } from './_util'
 
 describe('server', () => {
   let network: TestNetwork
@@ -20,56 +19,45 @@ describe('server', () => {
   })
 
   it('preserves 404s.', async () => {
-    const promise = axios.get(`${ozone.url}/unknown`)
-    await expect(promise).rejects.toThrow('failed with status code 404')
+    const response = await fetch(`${ozone.url}/unknown`)
+    expect(response.status).toEqual(404)
   })
 
   it('error handler turns unknown errors into 500s.', async () => {
     const app = express()
-    app.get('/oops', () => {
-      throw new Error('Oops!')
-    })
-    app.use(errorHandler)
-    const srv = app.listen()
-    const port = (srv.address() as AddressInfo).port
-    const promise = axios.get(`http://localhost:${port}/oops`)
-    await expect(promise).rejects.toThrow('failed with status code 500')
-    srv.close()
+      .get('/oops', () => {
+        throw new Error('Oops!')
+      })
+      .use(errorHandler)
+
+    const { origin, stop } = await startServer(app)
     try {
-      await promise
-    } catch (err: unknown) {
-      const axiosError = err as AxiosError
-      expect(axiosError.response?.status).toEqual(500)
-      expect(axiosError.response?.data).toEqual({
+      const response = await fetch(new URL(`/oops`, origin))
+      expect(response.status).toEqual(500)
+      await expect(response.json()).resolves.toEqual({
         error: 'InternalServerError',
         message: 'Internal Server Error',
       })
+    } finally {
+      await stop()
     }
   })
 
   it('healthcheck succeeds when database is available.', async () => {
-    const { data, status } = await axios.get(`${ozone.url}/xrpc/_health`)
-    expect(status).toEqual(200)
-    expect(data).toEqual({ version: '0.0.0' })
+    const response = await fetch(`${network.bsky.url}/xrpc/_health`)
+    expect(response.status).toEqual(200)
+    await expect(response.json()).resolves.toEqual({ version: 'unknown' })
   })
 
   it('healthcheck fails when database is unavailable.', async () => {
     // destroy sequencer to release connection that would prevent the db from closing
     await ozone.ctx.sequencer.destroy()
     await ozone.ctx.db.close()
-    let error: AxiosError
-    try {
-      await axios.get(`${ozone.url}/xrpc/_health`)
-      throw new Error('Healthcheck should have failed')
-    } catch (err) {
-      if (axios.isAxiosError(err)) {
-        error = err
-      } else {
-        throw err
-      }
-    }
-    expect(error.response?.status).toEqual(503)
-    expect(error.response?.data).toEqual({
+
+    const res = await fetch(`${ozone.url}/xrpc/_health`)
+
+    expect(res.status).toEqual(503)
+    await expect(res.json()).resolves.toEqual({
       version: '0.0.0',
       error: 'Service Unavailable',
     })