+++ packages/internal/fetch-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto-labs/fetch-node
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`622654672`](https://github.com/bluesky-social/atproto/commit/6226546725d1bb0375e3c9e0d71af173e8253c4f)]:
+  - @atproto-labs/fetch@0.1.2
+
 ## 0.1.3
 
 ### Patch Changes